fix: backups fixtures available for major version upgrades
