--- tests/integration/mongodb/backups/conftest.py
@@ -1,172 +1,0 @@
-import base64
-import dataclasses
-import json
-import logging
-import os
-import subprocess
-import time
-
-import boto3
-import botocore.exceptions
-import pytest
-
-
-@dataclasses.dataclass(frozen=True)
-class ConnectionInformation:
-    access_key_id: str
-    secret_access_key: str
-    bucket: str
-
-
-@pytest.fixture(scope="session")
-def microceph() -> ConnectionInformation:
-    """Deploy microceph with rados-gateway and provide the credentials to access it."""
-    if not os.environ.get("CI") == "true":
-        raise Exception("Not running on CI. Skipping microceph installation. ")
-    logger.info("Setting up microceph")
-
-    # socket.gethostbyname() might return `127.0.0.1`, which does not work from
-    # inside lxd container
-    host_ip = (
-        subprocess.run(["hostname", "-I"], capture_output=True, check=True, encoding="utf-8")
-        .stdout.strip()
-        .split()[0]
-    )
-
-    subprocess.run(["sudo", "snap", "install", "microceph"], check=True)
-    subprocess.run(["sudo", "microceph", "cluster", "bootstrap"], check=True)
-    subprocess.run(["sudo", "microceph", "disk", "add", "loop,4G,3"], check=True)
-    subprocess.run(
-        [
-            "openssl",
-            "req",
-            "-x509",
-            "-newkey",
-            "rsa:4096",
-            "-keyout",
-            "key.pem",
-            "-out",
-            "cert.pem",
-            "-sha256",
-            "-days",
-            "365",
-            "-nodes",
-            "-subj",
-            f"/CN={host_ip}",
-            "-addext",
-            f"subjectAltName=IP:{host_ip}",
-        ],
-        check=True,
-    )
-
-    with open("cert.pem", "rb") as cert_file:
-        cert = cert_file.read()
-        cert_encoded = base64.b64encode(cert)
-
-    with open("key.pem", "rb") as key_file:
-        key = key_file.read()
-        key_encoded = base64.b64encode(key)
-
-    subprocess.run(
-        [
-            "sudo",
-            "microceph",
-            "enable",
-            "rgw",
-            "--ssl-port",
-            "445",
-            "--ssl-certificate",
-            cert_encoded,
-            "--ssl-private-key",
-            key_encoded,
-        ],
-        check=True,
-    )
-    output = subprocess.run(
-        [
-            "sudo",
-            "microceph.radosgw-admin",
-            "user",
-            "create",
-            "--uid",
-            "test",
-            "--display-name",
-            "test",
-        ],
-        capture_output=True,
-        check=True,
-        encoding="utf-8",
-    ).stdout
-    key = json.loads(output)["keys"][0]
-    key_id = key["access_key"]
-    secret_key = key["secret_key"]
-    logger.info("Creating microceph bucket")
-    for attempt in range(3):
-        try:
-            boto3.client(
-                "s3",
-                endpoint_url=f"https://{host_ip}:445",
-                aws_access_key_id=key_id,
-                aws_secret_access_key=secret_key,
-                verify="cert.pem",
-            ).create_bucket(Bucket=_BUCKET)
-        except botocore.exceptions.EndpointConnectionError:
-            if attempt == 2:
-                raise
-            # microceph is not ready yet
-            logger.info("Unable to connect to microceph via S3. Retrying")
-            time.sleep(1)
-        else:
-            break
-    logger.info("Set up microceph")
-    return ConnectionInformation(key_id, secret_key, _BUCKET)
-
-
-_BUCKET = "testbucket"
-logger = logging.getLogger(__name__)
-
-
-@pytest.fixture(scope="session")
-def storage_config(microceph: ConnectionInformation) -> dict[str, str]:
-    """Provide the configuration required by s3-integrator."""
-    # socket.gethostbyname() might return `127.0.0.1`, which does not work from
-    # inside lxd container
-    host_ip = (
-        subprocess.run(["hostname", "-I"], capture_output=True, check=True, encoding="utf-8")
-        .stdout.strip()
-        .split()[0]
-    )
-
-    with open("cert.pem", "rb") as cert_file:
-        cert = cert_file.read()
-        cert_encoded = base64.b64encode(cert).decode("utf-8")
-
-    return {
-        "endpoint": f"https://{host_ip}:445",
-        "bucket": microceph.bucket,
-        "path": "mongodb-backups",
-        "region": "",
-        "tls-ca-chain": cert_encoded,
-    }
-
-
-@pytest.fixture(scope="session")
-def storage_credentials(microceph: ConnectionInformation) -> dict[str, str]:
-    """Provide the access-credentials required by s3-integrator."""
-    return {
-        "access-key": microceph.access_key_id,
-        "secret-key": microceph.secret_access_key,
-    }
-
-
-@pytest.fixture(scope="function")
-def s3_bucket(storage_credentials, storage_config) -> None:
-    """Provide a storage bucket on the deployed microceph instance."""
-    session = boto3.Session(
-        aws_access_key_id=storage_credentials["access-key"],
-        aws_secret_access_key=storage_credentials["secret-key"],
-        region_name=storage_config["region"] if storage_config["region"] else None,
-    )
-    s3 = session.resource("s3", endpoint_url=storage_config["endpoint"], verify="cert.pem")
-    bucket = s3.Bucket(storage_config["bucket"])
-    yield bucket
